refactor(repo C2): extract pure path utilities to engine/repository/paths.ts
Moves homedir, isAbsolutePath, normalizePathSeparators,
getRelativePathFromPrefix, resolve, enableProductionMode +
_productionMode state, getDefaultDbPath, getPwd, getRealPath
into engine/repository/paths.ts.

Virtual-path utilities (parseVirtualPath, resolveVirtualPath, etc.)
remain in engine/repository.ts because resolveVirtualPath depends on
getCollectionByName, which moves in a later cluster. Splitting them
now would create an ESM cycle.

engine/repository.ts: 4974 → 4772 lines.
Build: clean. Tests: 970/970 pass.
Regression bench: improvementPercent 91.74 (matches baseline).

Spec: docs/superpowers/specs/2026-05-20-kindx-strategic-refactor-program-design.md §5 cluster C2
